PYCBC-458: Fix Clang and Python warnings during installation
Motivation
----------
There are various Clang compiler warnings that come up during
installation, which should be addresed. Additionally, the
acouchbase package is included in Python<3.4 packages,
but this package relies on asyncio which is only in Python 3.4
and above. Finally, the package also uses the 'yield from'
command which was only introduced in Python 3.3. Therefore
the acouchbase package should not be included in Python<3.4
packages.

Modifications
-------------
- Fixed various Clang warnings.
- Change Ping code to only fill the 'details' field when an
  error has occurred, matching the behaviour of the JSON
  code from LCB. Also use the svc->rc field rather than
  svc->status, as for LCB. Adjust JSON schema in
  tests to allow for optional 'details' field.
Results
-------
All known warnings/errors eradicated. acouchbase package still
present in Python>=3.4, with all tests passing.

Change-Id: I4ce792fd8e0a1021c2c7fcbff5dfefd9393b5cd0