PYCBC-562: Move to lcb3 including collections support PYCBC-564: Add synchronous durability operations
This change migrates the codebase to the LCB V4 API
and adds the basis for Synchronous Replication and
Collections support.

It supports both PYCBC V2 and PYCBC V3. The
idea is to wrap the LCB V3 API and the features it supports
in the V4 API, or V4-style functions/macros.

Everything is, as far as possible,
written against the V4 API or an abstraction thereof.

Mostly the V4 API changes
affect:

1. creational logic:
 wrapped using CMDSCOPE_NG etc, as defined in pycbc.h

2. 'shared' attributes of structures.
 The backport of the V4 API (see lcb_v4_backport.h) to
 LCB API V3 replicates these interface changes as far
 as possible, so oneunified set of code can be used
 with both LCB API V3 and LCB API V4.

 There are some mutually exclusive features between
 LCB API V3 and LCB API V4
 (e.g. standalone ENDURE/OBSERVE commands)
 that are currently implemented via conditional compilation.

SDKJENKINS_PIP_INSTALL
SDKJENKINS_LINUX
PYCBC_BYPASS_V3_FAILURES

Change-Id: I6053c26cf8cc9ade78b63ec89bfd425d1ae35e60